Moved elf parsing/scattering to elf.c, WIP sim
