+++ arch.mk
@@ -1125,7 +1125,9 @@ ifeq ($(TARGET),sim)
   LD_END_GROUP=
   BOOT_IMG=test-app/image.elf
   CFLAGS+=-DARCH_SIM
-  CFLAGS+=-DWOLFBOOT_USE_STDLIBC
+  ifneq ($(ELF_SCATTERED),1)
+    CFLAGS+=-DWOLFBOOT_USE_STDLIBC
+  endif
   ifeq ($(FORCE_32BIT),1)
     CFLAGS+=-m32
     LDFLAGS+=-m32
