+++ options.mk
@@ -786,6 +786,10 @@ ifeq ($(ELF),1)
   ifneq ($(DEBUG_ELF),)
     CFLAGS+=-DDEBUG_ELF=$(DEBUG_ELF)
   endif
+  ifeq ($(ELF_SCATTERED),1)
+    CFLAGS+=-D"ELF_SCATTERED=1"
+  endif
+
 endif
 
 ifeq ($(MULTIBOOT2),1)
